fix: match product_id/invoice_id FK columns to the real (signed int) live schema
Root cause of the failed production deploy: product.id and invoice.id
are plain SIGNED int on this app's actual live schema (confirmed via
SHOW CREATE TABLE on production) - they predate Laravel's bigint id()
convention, and every one of the app's 11+ original 2026-07-10
migrations already treats product_id/invoice_id as plain integer().
Two migration files broke that discipline and used foreignId()/
unsignedBigInteger() instead, which MySQL refuses to constrain against
a mismatched-type primary key (errno 150 "incorrectly formed", or 3780
"incompatible" on a from-scratch install where product.id was still
bigint per its own migration's id() call).

Fixes, verified against a real local MySQL both ways - a fully fresh
migrate from zero, and a replica of production's actual current
partially-applied state (quotation_items.product_id already added as
the wrong type from the failed deploy, no constraint):

- create_product_table / create_invoice_table: id() column changed to
  integer('id', true), matching the real live column exactly. Already
  run on production; this only fixes future fresh installs.
- client_machines / spare_part_requests: product_id/invoice_id
  columns changed from unsignedBigInteger() to integer(), matching the
  established convention. Also already run on production - fresh-
  install correctness only.
- add_product_link_to_quotation_items_table (job_materials the same):
  changed to integer('product_id'), and made self-healing for
  production's current state - drops the wrong-typed column the failed
  deploy attempt left behind before re-adding it correctly, so the
  next migrate run repairs itself rather than erroring on a duplicate
  column.

Full local suite still green (547 passed) - SQLite never enforces FK
type matching, which is exactly why this shipped unnoticed until a
real MySQL deploy hit it.

+++ database/migrations/2026_07_26_111401_create_client_machines_table.php
@@ -14,12 +14,14 @@ return new class extends Migration
         Schema::create('client_machines', function (Blueprint $table) {
             $table->id();
             $table->foreignId('client_account_id')->constrained('client_accounts')->cascadeOnDelete();
-            // product/invoice.id are plain signed `int` on this app's live schema
-            // (predates Laravel's bigint id() convention) - see jobs table
-            // migration for the same discipline. foreignId() would create a
-            // column type MySQL refuses as FK-incompatible with the real tables.
-            $table->unsignedBigInteger('product_id');
-            $table->unsignedBigInteger('invoice_id')->nullable();
+            // product/invoice.id are plain SIGNED `int` on this app's live
+            // schema (predates Laravel's bigint id() convention - see
+            // create_product_table/create_invoice_table). unsignedBigInteger()
+            // (and foreignId()) create a column type MySQL refuses as
+            // FK-incompatible with the real tables - confirmed live via the
+            // identical errno 150/3780 on the quotation_items migration.
+            $table->integer('product_id');
+            $table->integer('invoice_id')->nullable();
             $table->string('serial_number');
             $table->date('installed_at')->nullable();
             $table->timestamps();
